# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
...
    AutomationDto:
      type: object
      properties:
        kind:
          type: string
        label:
          type: string
        state:
          type: string
        detail:
          type: string
        bumpId:
          type: string
        runIds:
          type: array
          items:
            type: string
        branch:
          type: string
        resultSha:
          type: string
        updatedAt:
          $ref: "#/components/schemas/Instant"
        failure:
          $ref: "#/components/schemas/FailureDto"
        reason:
          type: string
...
    DependencyDto:
      type: object
      properties:
        ecosystem:
          type: string
        name:
          type: string
        latest:
          type: string
        checkedAt:
          $ref: "#/components/schemas/Instant"
        error:
          type: string
        pins:
          type: array
          items:
            $ref: "#/components/schemas/DependencyPinDto"
    DependencyPinDto:
      type: object
      properties:
        repository:
          type: string
        version:
          type: string
        manifestPath:
          type: string
        pending:
          type: boolean
    DependentDto:
      type: object
      properties:
        artifactEcosystem:
          type: string
        artifactName:
          type: string
        artifactVersion:
          type: string
        repository:
          type: string
        embeddedVersion:
          type: string
        direct:
          type: boolean
        occurredAt:
          $ref: "#/components/schemas/Instant"
        sbomStatus:
          type: string
    DependentsDto:
      type: object
      properties:
        ecosystem:
          type: string
        name:
          type: string
        latest:
          type: string
        dependents:
          type: array
          items:
            $ref: "#/components/schemas/DependentDto"
...
    FailureDto:
      type: object
      properties:
        stepIndex:
          type: integer
          format: int32
        image:
          type: string
        exitCode:
          type: integer
          format: int32
        excerpt:
          type: string
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
...
    ReleaseRequestAutomationsDto:
      type: object
      properties:
        requestId:
          type: string
        foldSha:
          type: string
        automations:
          type: array
          items:
            $ref: "#/components/schemas/AutomationDto"
...
    RunRequest:
      type: object
      properties:
        workItem:
          type: string
        repository:
          type: string
...
    TriggerRequest:
      type: object
      properties:
        repository:
          type: string
        foldSha:
          type: string
        previousFoldSha:
          type: string
        changedSincePrevious:
          type: array
          items:
            type: string
        sourceBranches:
          type: array
          items:
            type: string
        workItem:
          type: string
        accepts:
          type: array
          items:
            type: string
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      description: Authentication
paths:
...
  /maintenance/api/dependencies:
    get:
      summary: "Dependencies matching a glob, with every pin of each"
      operationId: listDependencies
      parameters:
      - name: kind
        in: query
        schema:
          type: string
      - name: name
        in: query
        schema:
          type: string
          default: '*'
      responses:
        "200":
          description: The dependencies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DependencyDto"
        "400":
          description: kind is neither INTERNAL nor EXTERNAL
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Dependency Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
  /maintenance/api/dependencies/dependents:
    get:
      summary: Every released artifact of ours that embeds this dependency
      operationId: listDependencyDependents
      parameters:
      - name: all
        in: query
        schema:
          type: boolean
          default: false
      - name: ecosystem
        in: query
        schema:
          type: string
      - name: name
        in: query
        schema:
          type: string
      responses:
        "200":
          description: The dependents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DependentsDto"
        "400":
          description: "Unknown ecosystem, or no name"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Dependency Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
...
  /maintenance/api/release-requests/{requestId}/automations:
    post:
      summary: Settle a release request's automations at one fold
      operationId: triggerReleaseRequestAutomations
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerRequest"
        required: true
      responses:
        "200":
          description: Every applicable kind's state at the fold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseRequestAutomationsDto"
        "400":
          description: "Not a request id, not a sha, or not a work item"
        "404":
          description: No such repository in the inventory yet; a scan of it is queued
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Release Request Automation Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
    get:
      summary: A release request's automations at one fold
      operationId: listReleaseRequestAutomations
      parameters:
      - name: requestId
        in: path
        required: true
        schema:
          type: string
      - name: foldSha
        in: query
        schema:
          type: string
      responses:
        "200":
          description: Every kind with a row at the fold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseRequestAutomationsDto"
        "400":
          description: "Not a request id, or not a sha"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Release Request Automation Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
  /maintenance/api/release-requests/{requestId}/automations/{kind}/runs:
    post:
      summary: Re-run one automation on a release request's current fold
      operationId: runReleaseRequestAutomation
      parameters:
      - name: kind
        in: path
        required: true
        schema:
          type: string
      - name: requestId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunRequest"
        required: true
      responses:
        "202":
          description: "Requested; poll GET /bumps/{id}"
        "400":
          description: "Not a request id, or not a work item"
        "404":
          description: "No such kind, or no such repository"
        "409":
          description: "The request takes no branch or has no fold, one is already\
            \ active for (request, kind), or bumping is disabled"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Release Request Automation Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:system
        - qits:agent
...
info:
  title: qits-platform-maintenance API
  version: 2026.1002.201448